Add dynamic hint text and inset text for job title
