+++ app/views/publishers/vacancies/review_banners/_closed.html.slim
@@ -2,7 +2,7 @@ h1.govuk-heading-l class="govuk-!-display-inline"
   = vacancy.job_title
   = govuk_tag(text: t("publishers.vacancies.show.heading_component.status_tag.closed"), html_attributes: { class: "vertical-align-middle govuk-!-margin-left-1" }, colour: "red")
 = govuk_inset_text do
-  = vacancy_review_heading_inset_text(vacancy, "closed")
+  = vacancy_review_form_heading_inset_text(vacancy, "closed")
   .govuk-button-group class="govuk-!-margin-top-4"
     - %w[view copy extend_closing_date].each do |action|
-      = vacancy_review_heading_action_link(vacancy, action) unless vacancy.legacy? && action == "extend_closing_date"
+      = vacancy_review_form_heading_action_link(vacancy, action) unless vacancy.legacy? && action == "extend_closing_date"
