+++ app/views/publishers/vacancies/review_banners/_incomplete_draft.html.slim
@@ -2,7 +2,7 @@ h1.govuk-heading-l class="govuk-!-display-inline"
   = vacancy.job_title
   = govuk_tag(text: t("publishers.vacancies.show.heading_component.status_tag.draft"), html_attributes: { class: "vertical-align-middle govuk-!-margin-left-1" }, colour: "grey")
 = govuk_inset_text do
-  = vacancy_review_heading_inset_text(vacancy, "incomplete_draft")
+  = vacancy_review_form_heading_inset_text(vacancy, "incomplete_draft")
   .govuk-button-group class="govuk-!-margin-top-4"
     - %w[complete copy delete].each do |action|
-      = vacancy_review_heading_action_link(vacancy, action) unless vacancy.legacy_draft? && action == "complete"
+      = vacancy_review_form_heading_action_link(vacancy, action) unless vacancy.legacy_draft? && action == "complete"
