+++ app/views/publishers/vacancies/review_banners/_published.html.slim
@@ -2,7 +2,7 @@ h1.govuk-heading-l class="govuk-!-display-inline"
   = vacancy.job_title
   = govuk_tag(text: t("publishers.vacancies.show.heading_component.status_tag.published"), html_attributes: { class: "vertical-align-middle govuk-!-margin-left-1" }, colour: "green")
 = govuk_inset_text do
-  = vacancy_review_heading_inset_text(vacancy, "published")
+  = vacancy_review_form_heading_inset_text(vacancy, "published")
   .govuk-button-group class="govuk-!-margin-top-4"
     - %w[view copy close_early extend_closing_date].each do |action|
-      = vacancy_review_heading_action_link(vacancy, action)
+      = vacancy_review_form_heading_action_link(vacancy, action)
